+++ gamma_dashboard/static/dashboard/js/app/src/api/constants.js
@@ -12,6 +12,10 @@ export const BADGE_LEADERBOARD_URLS = (badgeSlug = '', courseId = '') => ({
   getInfo: `${BASE_URL}leaderboard/badge/${badgeSlug}${courseId ? `/${courseId}` : ''}`,
 });
 
+export const COURSE_LEADERBOARD_URLS = (courseId = '') => ({
+  getInfo: `${BASE_URL}leaderboard/course/${courseId}`,
+});
+
 export const DASHBOARD_URLS = {
   getGameProfile: `${BASE_URL}game-profile/`,
   selectUserAvatarSet: () => USER_AVATAR_CONFIG_URL,